% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{rsmetrics_traffic_transmitted}
\alias{rsmetrics_traffic_transmitted}
\title{Measure the Network Traffic Transmitted Out from an Application}
\usage{
rsmetrics_traffic_transmitted(from = "31d", until = "0d", interval = "1d")
}
\arguments{
\item{from, until}{Date range starting timestamp (Unix timestamp or
relative time delta such as "2d" or "3w").}

\item{interval}{Summarization interval. Data points at intervals less then
this will be grouped. (Relative time delta e.g. "120s" or "1h" or "30d").}
}
\value{
A tibble containing the metrics with the values scaled to megabytes.
}
\description{
This function retrieves and processes the metrics related to the network
traffic transmitted out from an application. The metrics are retrieved using
the \code{rsconnect::showMetrics} function and the results are converted to a
tibble and then returned in megabytes.
}
\examples{
\dontrun{
  # Return the average hourly network traffic per week over the past three
  # months
  traffic_data <- rsmetrics_traffic_transmitted(from = "90d", interval = "1w")
}
}
